Extract isNewGauge to utils
