+++ src/pages/vebal-voting.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import MultiVoting from '@/components/contextual/pages/vebal/LMVoting/MultiVoting.vue';
+import MultiVoting from '@/components/contextual/pages/vebal/MultiVoting/MultiVoting.vue';
 import { provideVoting } from '@/components/contextual/pages/vebal/providers/voting.provider';
 
 provideVoting();
